Backend/requirements.txt
crewai
langchain-groq
PyYAML
python-dotenv
python-socketio
fastapi
uvicorn[standard]
pydantic
pandas
numpy
pyexiv2
hachoir
requests
redis
pytz
geopy